requirements.txt
streamlit
pandas
google-generativeai
langchain
langchain-google-genai
openpyxl
